Command-line interface overhaul
The command-line interface (`cli.py`) has been heavily refactored.

Functionally speaking, the command-line now records the filenames
and module names of all inlined inputs for a Python -> Notebook
conversion. This information is propagated all the way to the
notebook, and can then be used to re-construct the Python source
tree from a Notebook